backend/src/scripts/requirements.txt
pandas
matplotlib
reportlab
python-dotenv
mysql-connector-python